[TASK] Ensure that the ControllerContext is available when needed
In older versions of the TYPO3\CMS\Fluid\Core\ViewHelper\AbstractViewhelper
the ControllerContext of Fluid/Extbase was set. This changed so we have
to do it on our own.

This is the only ViewHelper in our project that needs this ControllerContext.
Therefore we do not put this snippet into an AbstractViewhelper.